Support WebAssembly and asm.js
This adds support for both WebAssembly and asm.js by abstracting the
worker thread into an ImmediateWorker and a ThreadWorker which uses the
ImmediateWorker. The ThreadWorker is exposed as the worker to be used on
every platform other than WebAssembly and asm.js. There the
ImmediateWorker itself is exposed instead, because threading isn't
supported on those platforms just yet.

Resolves #75