Fix MCU reference in odd decoding order
This addresses a potential panic due to index errors. When a component
is finished then it can be put into a queue. This references the MCU row
coefficients, some structure filled for each _finished_ component.

However, the list for these MCU row coefficients is allocated lazily.
Previously we would fill it with an instance for each _finished_
component but then used the index of the component (in the list of _all_
components) to access it. This could cause wrong indices or a
past-the-end index and subsequent panic.