Restructure multithreading
Redefine the Worker trait, which allows the chosen worker to create a
new scope. This is relevant for a newly created (and installed) rayon
worker.

The rayon thread pool is now local to the decoding step. This fixes an
issue where improper task scheduling would deadlock decoding. It's not
clear how the intended task scheduling can be reliably achieved without
the guarantee of having at least a second, free, worker thread.